* Avoid a compiler warning

+++ plugins/e131/e131/PreamblePacker.cpp
@@ -77,7 +77,8 @@ const uint8_t *PreamblePacker::Pack(const PDUBlock<PDU> &pdu_block,
   if (!m_send_buffer)
     Init();
 
-  unsigned int size = MAX_DATAGRAM_SIZE - sizeof(ACN_HEADER);
+  unsigned int size = MAX_DATAGRAM_SIZE -
+      static_cast<unsigned int>(sizeof(ACN_HEADER));
   if (!pdu_block.Pack(m_send_buffer + sizeof(ACN_HEADER), &size)) {
     OLA_WARN << "Failed to pack E1.31 PDU";
     return NULL;
